Retrieves the distance and time of travel between two Latitude/Longitude coordinates.
It first decodes the coordinates.json and gets the initial and final
coordinates then it translates it into a URL which is sent off to Google
Maps server to be translated and given a result in JSON format. Which is
then decoded and the distance and time intervals are printed onto the
console.